Refactored solve to Use Nodes Instead of Integers
The solve method is the main method that kicks off the chain analysis.
Working with nodes instead of integers makes it a lot easier to send the
objects for analysis and creation. The only integers that remain are the
ones returned by even/odd and are immediately used for creating a new node.